Finalised .txt loading
Includes the ability to guess the delimiter (called escape character, for now) by entering nothing. It will assume it's either a \t or any number of " " characters.

If you know what your character is, you  should manually give it to the code.

Todo: strip spaces if there exists any to the left or right of the escape character