Fixed NewDict logic in DataDir
DataDir will now check to see if there are missing integers in the sequential list - and as a result, will also reorganise the list if it deems it necessary! This is done so that you can delete many items inside of a list, and still have it organised when you save it later!