Added LDI_Data_Import
This function will eventually replace doing most of your importing, but will REQUIRE that you use Get_FileList, and that you have a txt file named with a matching start string to all of your files. This .txt file can be empty, but is the basis of multi-import.
Modify this code to add more data if your monitors do not have it. Currently has problems on transmission monitors, will be fixed soon